feat(ui): merge arrange tools and changelog
